pyproject.toml
[project]
name = "receipt-voiceover"
version = "0.1.0"
description = "レシート画像を読み取り、VOICEVOXで音声で読み上げるプログラムです。"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"dotenv-flow>=0.4.0",
	"google-generativeai>=0.8.3",
	"pillow>=11.1.0",
	"pyaudio>=0.2.14",
	"requests>=2.32.3",
]

[tool.ruff.format]
indent-style = "tab"
